Reimplement rebase functionality correctly.
It was previously implemented naively: fully parse the analysis file,
then rewrite lines as they're written out.

It was also implemented too generally: allowing any number of rebasings,
whereas we only actually need a single rewrite (of pants home to/from a
placeholder) and, optionally, a single filter (of any lines containing
paths under jvm home).

This new implementation is much more selective and much more efficient:

- Each section now declares whether its lines need rewrite or filter or both.
  It also declares whether the entire line needs to be examined, or whether
  it's enough to just match at the start of the line.
- Instead of fully parsing the file, which is slow, rebasing now just
  rips over the lines and acts on them as text.

Also split ZincAnalysisElement into its own file, and its subclasses into
their own file, as zinc_analysis.py was getting quite huge.
